+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Add file_name to BasePersitentFile load
 - Add get_latest_response_by_params to session recorder
+- Add 'optional settings' config option
 - Remove individual ttl for session recorder entry
 - Rename filename to file_name
 - Use enum for internal custom ids